Crate aster_bigtcp

bigtcp is a crate that wraps [smoltcp].

[smoltcp] is designed for embedded systems where the number of sockets is always small. It turns out that such a design cannot satisfy the need to implement the network stack of a general-purpose OS kernel, in terms of flexibility and efficiency.

The short-term goal of bigtcp is to reuse the powerful TCP implementation of smoltcp, while reimplementing Ethernet and IP protocols to increase the flexibility and performance of packet dispatching.

Modules

boolean_value

device

errors

ext

iface

socket

socket_table

This module defines the socket table, which manages all TCP and UDP sockets, for efficiently inserting, looking up, and removing sockets.

time

wire

Macros

define_boolean_value

Defines a struct representing a boolean value.