aster_bigtcp::device

Trait AnyNetworkDevice

pub trait AnyNetworkDevice:
    Send
    + Sync
    + Any
    + Debug {
    // Required methods
    fn mac_addr(&self) -> EthernetAddress;
    fn capabilities(&self) -> DeviceCapabilities;
    fn can_receive(&self) -> bool;
    fn can_send(&self) -> bool;
    fn receive(&mut self) -> Result<RxPacket<LinkLayer>, NetError>;
    fn send(&mut self, packet: TxPacket<LinkLayer>) -> Result<(), NetError>;
    fn alloc_tx_buffer(payload_len: usize) -> Result<FreshTxPacket, Error>
       where Self: Sized;
    fn free_processed_tx_buffers(&mut self);
    fn notify_poll_end(&mut self);
}

Required Methods

fn mac_addr(&self) -> EthernetAddress

fn capabilities(&self) -> DeviceCapabilities

fn can_receive(&self) -> bool

fn can_send(&self) -> bool

fn receive(&mut self) -> Result<RxPacket<LinkLayer>, NetError>

Receives a link-layer packet from the network.

If no packet is ready, this method returns NetError::NotReady.

fn send(&mut self, packet: TxPacket<LinkLayer>) -> Result<(), NetError>

Sends a link-layer packet to the network.

It is recommended that the packet is allocated using Self::alloc_tx_buffer to improve performance (e.g., by reducing DMA overhead via a pool). However, this is not a strict requirement, and the method must work regardless of how the packet is allocated.

fn alloc_tx_buffer(payload_len: usize) -> Result<FreshTxPacket, Error>
where Self: Sized,

Allocates a packet buffer for transmission through this device.

fn free_processed_tx_buffers(&mut self)

Frees processed transmit buffers.

fn notify_poll_end(&mut self)

Notifies the device driver that a polling operation has ended.

The driver can assume that the device remains protected by a poll lock for the entire duration of polling, so two polling operations cannot happen simultaneously.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".