aster_bigtcp::device

Trait WithDevice

pub trait WithDevice: Send + Sync {
    type Device: Device + ?Sized;

    // Required method
    fn with<F, R>(&self, f: F) -> R
       where F: FnOnce(&mut Self::Device) -> R;
}

A trait that allows to obtain a mutable reference of Device.

A Device is usually protected by a lock (e.g., a spin lock or a mutex), and it may be stored behind a shared type (e.g., an Arc). This property abstracts this fact by providing a method that the caller can use to get the mutable reference without worrying about how the reference is obtained.

Required Associated Types

type Device: Device + ?Sized

Required Methods

fn with<F, R>(&self, f: F) -> R
where F: FnOnce(&mut Self::Device) -> R,

Calls the closure with a mutable reference of Device.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.