aster_bigtcp::ext

Trait Ext

pub trait Ext {
    type ScheduleNextPoll: ScheduleNextPoll;
    type TcpEventObserver: SocketEventObserver + Clone;
    type UdpEventObserver: SocketEventObserver;
}

Extension to be implemented by users of this crate.

This should be implemented on an empty type that carries no data, since the type will never actually be instantiated.

The purpose of having this trait is to allow users of this crate to inject multiple types without the hassle of writing multiple trait bounds, which can be achieved by using the types associated with this trait.

Required Associated Types

type ScheduleNextPoll: ScheduleNextPoll

The type for ifaces to schedule the next poll.

type TcpEventObserver: SocketEventObserver + Clone

The type for TCP sockets to observe events.

type UdpEventObserver: SocketEventObserver

The type for UDP sockets to observe events.