aster_bigtcp::iface

Trait Iface

pub trait Iface<E>:
    IfaceInternal<E>
    + Send
    + Sync {
    // Required methods
    fn poll(&self);
    fn mtu(&self) -> usize;
}

A network interface.

A network interface (abbreviated as iface) is a hardware or software component that connects a computer to a network. Network interfaces can be physical components like Ethernet ports or wireless adapters. They can also be virtual interfaces created by software, such as virtual private network (VPN) connections.

Required Methods

fn poll(&self)

Transmits or receives packets queued in the iface, and updates socket status accordingly.

fn mtu(&self) -> usize

Returns the maximum transmission unit.

Implementations

impl<E: Ext> dyn Iface<E>

pub fn bind( self: &Arc<Self>, config: BindPortConfig, ) -> Result<BoundPort<E>, BindError>

Binds a socket to the iface.

After binding the socket to the iface, the iface will handle all packets to and from the socket.

If BindPortConfig::Ephemeral is specified, the iface will pick up an ephemeral port for the socket.

FIXME: The reason for binding the socket and the iface together is because there are limitations inside smoltcp. See discussion at https://github.com/smoltcp-rs/smoltcp/issues/779.

pub fn index(&self) -> u32

Returns the interface index.

pub fn name(&self) -> &str

Gets the name of the iface.

In Linux, the name is usually the driver name followed by a unit number.

pub fn type_(&self) -> InterfaceType

Returns the interface type.

pub fn flags(&self) -> InterfaceFlags

Returns the interface flags.

pub fn ipv4_addr(&self) -> Option<Ipv4Address>

Gets the IPv4 address of the iface, if any.

pub fn prefix_len(&self) -> Option<u8>

Retrieves the prefix length of the interface’s IPv4 address.

Both Self::ipv4_addr and this method will either return Some(_) or both will return None.

pub fn broadcast_addr(&self) -> Option<Ipv4Address>

Gets the broadcast address of the iface, if any.

pub fn sched_poll(&self) -> &E::ScheduleNextPoll

Returns a reference to the associated ScheduleNextPoll.