aster_bigtcp

Module packet

Network packet abstractions.

Network packets consist of headers from multiple layers and the actual payload. When sending or receiving a packet, the headers are added or removed as the packet goes through the network stack.

We use TxPacket and RxPacket, plus a marker that specifies the layer (DeviceLayer, LinkLayer, NetworkLayer, TransportLayer, and ApplicationLayer), to denote a packet that is being processed by that that network layer.

Here is an illustration of TxPacket<TransportLayer> (the part above) and RxPacket<NetworkLayer> (the part below):

                                  TxPacket<TransportLayer>
                                            |
<--------- Headers (to be filled) --------> | <------- Payload (filled) ---------->
                                            |
                                            v
+--------------+------------+---------------+-----------------+-------------------+
| Device Layer | Link Layer | Network Layer | Transport Layer | Application Layer |
+--------------+------------+---------------+-----------------+-------------------+
                            ^
                            |
<--- Headers (consumed) --> | <----------- Payload (to be consumed) -------------->
                            |
                        RxPacket<NetworkLayer>

Structs

FreshTxPacket

A fresh, empty TX packet that has just been allocated.

RxBuffer

An RX buffer, which is DMA-mapped and can be converted to an RX packet later.

RxPacket

An RX packet.

TxBuffer

A TX buffer, which is a DMA-mapped TX packet.

TxPacket

A TX packet.

TxPacketBuilder

A TX packet builder that fills some payload into the packet.

Enums

ApplicationLayer

DeviceLayer

LinkLayer

NetworkLayer

TransportLayer

Traits

Layer

A network layer.