aster_bigtcp::packet

Struct TxPacket

pub struct TxPacket<L>(/* private fields */);

A TX packet.

The life cycle of a TX packet contains the following phases:

Allocation: FreshTxPacket::alloc and FreshTxPacket::alloc_from_pool.

Filling the payload: TxPacketBuilder::append_writer and TxPacketBuilder::commit.

Filling the headers: TxPacket::prepend_writer and TxPacket::pack.

Creating DMA buffer: TxPacket::map_dma and TxPacket::to_dma_buffer.

Device transmission: RxBuffer, which implements [HasDaddr].

Implementations

impl<L> TxPacket<L>

pub fn reader(&self) -> VmReader<'_, Infallible>

Returns a reader, starting at the L layer.

pub fn reader_with_header(&self, header_len: usize) -> VmReader<'_, Infallible>

Returns a reader, starting at header_len bytes before the L layer.

Callers should have used Self::prepend_writer to fill at least header_len bytes before the L layer.

Panics

This method will panic if header_len exceeds the allocated header size.

pub fn len(&self) -> usize

Returns the number of bytes at the L layer.

pub fn prepend_writer(&mut self, header_len: usize) -> VmWriter<'_, Infallible>

Returns a writer that can prepend a header at header_len bytes before the L layer.

Panics

This method will panic if header_len exceeds the allocated header size.

pub fn reset_to_fresh(self) -> FreshTxPacket

Resets the packet back to a fresh, empty TX packet.

Panics

This method will panic if the allocated size is less than the default value of the allocated header size, which is ApplicationLayer::HEAD_ROOM_SIZE.

pub fn to_rx_packet(self) -> Option<RxPacket<L>>

Converts to an RX packet.

If the TX packet is associated with a DMA buffer (i.e., if it was oringinally allocated via FreshTxPacket::alloc_from_pool instead FreshTxPacket::alloc), this method will return None.

impl<L: Layer> TxPacket<L>

pub fn prepend_and_pack<T: Pod>(self, header: &T) -> TxPacket<L::Preceding>

Prepends a header before the L layer and goes down to the L::Preceding layer.

This operation is equivalent to writing the header to Self::prepend_writer and then adjusting the layer with Self::pack.

Panics

This method will panic if size_of::<T>() exceeds the allocated header size.

pub fn pack(self, header_len: usize) -> TxPacket<L::Preceding>

Goes down to the L::Preceding layer, assuming that a header whose length is header_len has been filled.

Panics

This method will panic if the total header length after the operation exceeds the allocated header size.

impl<L: Layer<Preceding = ()>> TxPacket<L>

pub fn map_dma(self, is_cache_coherent: bool) -> Result<TxBuffer>

Ensures that the DMA buffer is up to date regarding the new bytes in the packet.

If the TX packet is not associated with a DMA buffer, this method will attempt to create one and associate the packet with it.

pub fn to_dma_buffer(self) -> Option<TxBuffer>

Tries to ensure that the DMA buffer is up to date regarding the new bytes in the packet.

If the TX packet is not associated with a DMA buffer (i.e., if it was oringinally allocated via FreshTxPacket::alloc instead FreshTxPacket::alloc_from_pool), this method will return None.