aster_bigtcp::packet

Trait Layer

pub trait Layer {
    type Preceding: Layer;

    const MAX_HEADER_SIZE: usize;
    const HEAD_ROOM_SIZE: usize = _;
}

A network layer.

Required Associated Constants

const MAX_HEADER_SIZE: usize

The maximum header size for all devices/protocols that we support in this layer.

Note that the value should be adjusted when supporting new devices/protocols that require large headers.

Provided Associated Constants

const HEAD_ROOM_SIZE: usize = _

The sum of the Layer::MAX_HEADER_SIZE for all layers preceding this one.

Required Associated Types

type Preceding: Layer

The network layer preceding to this layer.

This should be specified as the unit type () for the first layer with no preceding layers.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".