aster_bigtcp

Module socket

Structs

NeedIfacePoll

Whether the iface needs to be polled

RawTcpOption

Socket options on a raw socket.

RawTcpSocketExt

SocketEvents

Socket events caused by the network.

Enums

ConnectState

Constants

TCP_RECV_BUF_LEN

TCP_SEND_BUF_LEN

UDP_RECV_PAYLOAD_LEN

UDP_SEND_PAYLOAD_LEN

Traits

RawTcpSetOption

A trait defines setting socket options on a raw socket.

SocketEventObserver

A observer that will be invoked whenever events occur on the socket.

Type Aliases

RawUdpSocket

TcpConnection

TcpListener

UdpSocket