aster_bigtcp::socket

Struct RawTcpSocketExt

pub struct RawTcpSocketExt<E: Ext> { /* private fields */ }

Implementations

impl<E: Ext> RawTcpSocketExt<E>

pub fn may_recv_new(&self) -> bool

Checks if the socket may receive any new data.

This is similar to [RawTcpSocket::may_recv]. However, this method checks if there can be new data. In other words, if there is already buffered data in the socket, [RawTcpSocket::may_recv] will always return true since it is possible to receive the buffered data, but this method may return false if the peer has closed its sending half (so no new data can come in).

pub fn is_closing(&self) -> bool

Checks if the socket is closing.

More specifically, we say a socket is closing if and only if it has sent its FIN packet but is still waiting for an ACK packet from the peer to acknowledge the FIN it sent.

pub fn is_recv_shut(&self) -> bool

Returns whether the receiving half of the socket is shut down.

This method will return true if and only if [TcpConnection::shut_recv] or [TcpConnection::close] is called.

pub fn is_rst_closed(&self) -> bool

Checks if the socket is closed by a RST packet.

Note that the flag is automatically cleared when it is read by [TcpConnection::clear_rst_closed], [TcpConnection::send], or [TcpConnection::recv].