aster_bigtcp::socket

Trait RawTcpSetOption

pub trait RawTcpSetOption {
    // Required methods
    fn set_keep_alive(&self, interval: Option<Duration>) -> NeedIfacePoll;
    fn set_nagle_enabled(&self, enabled: bool);
}

A trait defines setting socket options on a raw socket.

Required Methods

fn set_keep_alive(&self, interval: Option<Duration>) -> NeedIfacePoll

Sets the keep alive interval.

Polling the iface may be required after this method succeeds.

fn set_nagle_enabled(&self, enabled: bool)

Enables or disables Nagle’s Algorithm.

Polling the iface is not required after this method succeeds.