aster_bigtcp::socket

Type Alias TcpConnection

pub type TcpConnection<E> = Socket<TcpConnectionInner<E>, E>;

Aliased Type

pub struct TcpConnection<E>(/* private fields */);

Implementations

impl<E: Ext> TcpConnection<E>

pub fn new_connect( bound: BoundPort<E>, remote_endpoint: IpEndpoint, option: &RawTcpOption, observer: E::TcpEventObserver, ) -> Result<Self, (BoundPort<E>, ConnectError)>

Connects to a remote endpoint.

Polling the iface is always required after this method succeeds.

pub fn connect_state(&self) -> ConnectState

Returns the state of the connecting procedure.

pub fn into_bound_port(self) -> Option<BoundPort<E>>

Converts back to the [BoundPort].

This method will succeed if the connection is fully closed and no network events can reach this connection. We guarantee that this method will always succeed if [Self::connect_state] returns [ConnectState::Refused].

pub fn send<F, R>(&self, f: F) -> Result<(R, NeedIfacePoll), SendError>
where F: FnOnce(&mut [u8]) -> (usize, R),

Sends some data.

Polling the iface may be required after this method succeeds.

pub fn recv<F, R>(&self, f: F) -> Result<(R, NeedIfacePoll), RecvError>
where F: FnOnce(&mut [u8]) -> (usize, R),

Receives some data.

Polling the iface may be required after this method succeeds.

pub fn clear_rst_closed(&self) -> bool

Checks if the socket is closed by a RST packet and clears the flag.

This flag is set when the socket is closed by a RST packet, and cleared when the connection reset error is reported via one of the [Self::send], [Self::recv], or this method.

pub fn shut_send(&self) -> bool

Shuts down the sending half of the connection.

This method will return false if the socket is in the CLOSED or TIME_WAIT state.

Polling the iface is always required after this method succeeds.

pub fn shut_recv(&self) -> bool

Shuts down the receiving half of the connection.

This method will return false if the socket is in the CLOSED or TIME_WAIT state.

Polling the iface is not required after this method succeeds.

pub fn close(&self)

Closes the connection.

Polling the iface is always required after this method succeeds.

Note that either this method or [Self::reset] must be called before dropping the TCP connection to avoid resource leakage.

pub fn reset(&self)

Resets the connection.

Polling the iface is always required after this method succeeds.

Note that either this method or [Self::close] must be called before dropping the TCP connection to avoid resource leakage.

pub fn raw_with<F, R>(&self, f: F) -> R
where F: FnOnce(&RawTcpSocketExt<E>) -> R,

Calls f with an immutable reference to the associated [RawTcpSocket].