aster_bigtcp::socket

Type Alias TcpListener

pub type TcpListener<E> = Socket<TcpListenerInner<E>, E>;

Aliased Type

pub struct TcpListener<E>(/* private fields */);

Implementations

impl<E: Ext> TcpListener<E>

pub fn new_listen( bound: BoundPort<E>, max_conn: usize, option: &RawTcpOption, observer: E::TcpEventObserver, ) -> Result<Self, (BoundPort<E>, ListenError)>

Listens at a specified endpoint.

Polling the iface is not required after this method succeeds.

pub fn accept(&self) -> Option<(TcpConnection<E>, IpEndpoint)>

Accepts a TCP connection.

Polling the iface is not required after this method succeeds.

pub fn can_accept(&self) -> bool

Returns whether there is a TCP connection to accept.

It’s the caller’s responsibility to deal with race conditions when using this method.

pub fn close(&self)

Closes the listener.

Polling the iface is always required after this method succeeds.

Note that this method must be called before dropping the TCP listener to avoid resource leakage.