aster_bigtcp::socket

Type Alias UdpSocket

pub type UdpSocket<E> = Socket<UdpSocketInner, E>;

Aliased Type

pub struct UdpSocket<E>(/* private fields */);

Implementations

impl<E: Ext> UdpSocket<E>

pub fn new_bind( bound: BoundPort<E>, observer: E::UdpEventObserver, ) -> Result<Self, (BoundPort<E>, BindError)>

Binds to a specified endpoint.

Polling the iface is not required after this method succeeds.

pub fn send<F, R>( &self, size: usize, meta: impl Into<UdpMetadata>, f: F, ) -> Result<R, SendError>
where F: FnOnce(&mut [u8]) -> R,

Sends some data.

Polling the iface is always required after this method succeeds.

pub fn recv<F, R>(&self, f: F) -> Result<R, RecvError>
where F: FnOnce(&[u8], UdpMetadata) -> R,

Receives some data.

Polling the iface is not required after this method succeeds.

pub fn raw_with<F, R>(&self, f: F) -> R
where F: FnOnce(&RawUdpSocket) -> R,

Calls f with an immutable reference to the associated [RawUdpSocket].