aster_bigtcp::time

Struct Duration

pub struct Duration { /* private fields */ }

A relative amount of time.

Implementations

impl Duration

pub const ZERO: Duration

pub const MAX: Duration

The longest possible duration we can encode.

pub const fn from_micros(micros: u64) -> Duration

Create a new Duration from a number of microseconds.

pub const fn from_millis(millis: u64) -> Duration

Create a new Duration from a number of milliseconds.

pub const fn from_secs(secs: u64) -> Duration

Create a new Duration from a number of seconds.

pub const fn millis(&self) -> u64

The fractional number of milliseconds in this Duration.

pub const fn micros(&self) -> u64

The fractional number of milliseconds in this Duration.

pub const fn secs(&self) -> u64

The number of whole seconds in this Duration.

pub const fn total_millis(&self) -> u64

The total number of milliseconds in this Duration.

pub const fn total_micros(&self) -> u64

The total number of microseconds in this Duration.