aster_bigtcp::wire

Enum IpAddress

pub enum IpAddress {
    Ipv4(Ipv4Addr),
}

An internetworking address.

Variants

Ipv4(Ipv4Addr)

An IPv4 address.

Implementations

impl Address

pub const fn v4(a0: u8, a1: u8, a2: u8, a3: u8) -> Address

Create an address wrapping an IPv4 address with the given octets.

pub const fn version(&self) -> Version

Return the protocol version.

pub fn is_unicast(&self) -> bool

Query whether the address is a valid unicast address.

pub const fn is_multicast(&self) -> bool

Query whether the address is a valid multicast address.

pub fn is_broadcast(&self) -> bool

Query whether the address is the broadcast address.

pub fn is_unspecified(&self) -> bool

Query whether the address falls into the โunspecifiedโ range.

pub fn prefix_len(&self) -> Option<u8>

If self is a CIDR-compatible subnet mask, return Some(prefix_len), where prefix_len is the number of leading zeroes. Return None otherwise.