aster_bigtcp::wire

Trait Ipv6AddressExt

pub trait Ipv6AddressExt {
    // Required methods
    fn from_bytes(data: &[u8]) -> Ipv6Addr;
    fn x_is_unicast(&self) -> bool;
    fn is_global_unicast(&self) -> bool;
    fn is_link_local(&self) -> bool;
    fn x_is_unique_local(&self) -> bool;
    fn mask(&self, mask: u8) -> [u8; 16];
    fn solicited_node(&self) -> Ipv6Addr;
    fn x_multicast_scope(&self) -> MulticastScope;
    fn prefix_len(&self) -> Option<u8>;
}

Required Methods

fn from_bytes(data: &[u8]) -> Ipv6Addr

Construct an IPv6 address from a sequence of octets, in big-endian.

Panics

The function panics if data is not sixteen octets long.

fn x_is_unicast(&self) -> bool

Query whether the IPv6 address is an unicast address.

x_ prefix is to avoid a collision with the still-unstable method in core::ip.

fn is_global_unicast(&self) -> bool

Query whether the IPv6 address is a global unicast address.

fn is_link_local(&self) -> bool

Query whether the IPv6 address is in the link-local scope.

fn x_is_unique_local(&self) -> bool

Query whether the IPv6 address is a Unique Local Address (ULA).

x_ prefix is to avoid a collision with the still-unstable method in core::ip.

fn mask(&self, mask: u8) -> [u8; 16]

Helper function used to mask an address given a prefix.

Panics

This function panics if mask is greater than 128.

fn solicited_node(&self) -> Ipv6Addr

The solicited node for the given unicast address.

Panics

This function panics if the given address is not unicast.

fn x_multicast_scope(&self) -> MulticastScope

Return the scope of the address.

x_ prefix is to avoid a collision with the still-unstable method in core::ip.

fn prefix_len(&self) -> Option<u8>

If self is a CIDR-compatible subnet mask, return Some(prefix_len), where prefix_len is the number of leading zeroes. Return None otherwise.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".